+++ src/main/java/org/chiclepad/backend/session/BadPasswordException.java
@@ -1,4 +1,4 @@
-package org.chiclepad.backend.usermanagement;
+package org.chiclepad.backend.session;
 
 public class BadPasswordException extends RuntimeException {
 }
