+++ src/main/java/org/chiclepad/backend/session/UserAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package org.chiclepad.backend.usermanagement;
+package org.chiclepad.backend.session;
 
 public class UserAlreadyExistsException extends RuntimeException {
 }
