+++ src/main/resources/org/chiclepad/frontend/jfx/style.css
@@ -164,6 +164,10 @@ tertiary: #3E5641
     -fx-background-color: #e0e0e0;
 }
 
+.jfx-list-cell:hover {
+    -fx-background-color: #d5d5d5;
+}
+
 .jfx-rippler {
     -jfx-rippler-fill: #3E5641;
 }
